requirements.txt
Django==1.10.4
Pygments==2.1.3
blessings==1.6
bpython==0.16
curtsies==0.2.11
djangorestframework==3.5.3
geojson==1.3.3
greenlet==0.4.10
requests==2.12.3
six==1.10.0
wcwidth==0.1.7

